Update TextInput file structure
